Moved autocomplete to other dir
